+++ services/src/main/java/org/openlearn/web/rest/PortfolioItemResource.java
@@ -1,11 +1,16 @@
 package org.openlearn.web.rest;
 
-import com.amazonaws.services.s3.model.S3ObjectSummary;
 import io.swagger.annotations.ApiParam;
+import org.openlearn.dto.FileInformationDTO;
 import org.openlearn.dto.PortfolioItemDTO;
 import org.openlearn.security.AuthoritiesConstants;
+import org.openlearn.security.SecurityUtils;
+import org.openlearn.service.FileInformationService;
 import org.openlearn.service.PortfolioItemService;
 import org.openlearn.service.StorageService;
+import org.openlearn.service.UserService;
+import org.openlearn.web.rest.errors.FileInformationNotFoundException;
+import org.openlearn.web.rest.errors.PortfolioItemNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -19,6 +24,7 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -38,9 +44,17 @@ public class PortfolioItemResource {
 
 	private final StorageService storageService;
 
-	public PortfolioItemResource(final PortfolioItemService portfolioItemService, final StorageService storageService) {
+	private final UserService userService;
+
+	private final FileInformationService fileInformationService;
+
+	public PortfolioItemResource(final PortfolioItemService portfolioItemService, final StorageService storageService,
+								 final UserService userService,
+								 final FileInformationService fileInformationService) {
 		this.portfolioItemService = portfolioItemService;
 		this.storageService = storageService;
+		this.userService = userService;
+		this.fileInformationService = fileInformationService;
 	}
 
 	/**
@@ -100,8 +114,14 @@ public class PortfolioItemResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity create(@RequestBody @Valid final PortfolioItemDTO portfolioItemDTO) throws URISyntaxException {
 		log.debug("POST request to create portfolio item : {}", portfolioItemDTO);
-		PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
-		return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		if (hasCreateUpdateDeleteAuthority(portfolioItemDTO)) {
+			PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
+			return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		} else {
+			log.info("User is not authorized to create portfolio items");
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
+
 	}
 
 	/**
@@ -115,8 +135,13 @@ public class PortfolioItemResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity update(@RequestBody @Valid final PortfolioItemDTO portfolioItemDTO) {
 		log.debug("PUT request to update portfolio item : {}", portfolioItemDTO);
-		PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
-		return ResponseEntity.ok(response);
+		if (hasCreateUpdateDeleteAuthority(portfolioItemDTO)) {
+			PortfolioItemDTO response = portfolioItemService.save(portfolioItemDTO);
+			return ResponseEntity.ok(response);
+		} else {
+			log.info("User is not authorized to update portfolio item: {}.", portfolioItemDTO.getId());
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 	/**
@@ -130,8 +155,14 @@ public class PortfolioItemResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity delete(@PathVariable final Long id) {
 		log.debug("DELETE request to delete portfolio item : {}", id);
-		portfolioItemService.delete(id);
-		return ResponseEntity.ok().build();
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(id);
+		if (hasCreateUpdateDeleteAuthority(portfolioItem)) {
+			portfolioItemService.delete(id);
+			return ResponseEntity.ok().build();
+		} else {
+			log.info("User is not authorized to delete portfolio item: {}.", id);
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 
@@ -144,57 +175,111 @@ public class PortfolioItemResource {
 	 */
 	@PostMapping(path="/{portfolioId}/upload")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
-	public String uploadCourseFile(@PathVariable final Long portfolioId,
-								   @RequestParam("file") MultipartFile file,
-								   RedirectAttributes redirectAttributes) throws URISyntaxException {
-		storageService.store(file, null, portfolioId);
-		redirectAttributes.addFlashAttribute("message",
-			"You successfully uploaded " + file.getOriginalFilename() + "!");
-
-		return "You successfully uploaded " + file.getOriginalFilename() + "!";
+	public ResponseEntity uploadCourseFile(@PathVariable final Long portfolioId,
+										@RequestParam("file") MultipartFile file,
+										RedirectAttributes redirectAttributes) throws URISyntaxException {
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		if (canUploadFilesToPortfolio(portfolioItem)) {
+			FileInformationDTO response = storageService.store(file, null, portfolioId);
+			redirectAttributes.addFlashAttribute("message",
+				"You successfully uploaded " + file.getOriginalFilename() + "!");
+			return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		} else {
+			log.info("User is not authorized to upload files for portfolio item: {}.", portfolioId);
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 	@GetMapping(path="/{portfolioId}/uploads")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
-	public ResponseEntity getUploads(@PathVariable final Long portfolioId) {
+	public ResponseEntity getUploads(@PathVariable final Long portfolioId, @ApiParam Pageable pageable) {
 		log.debug("GET request to get course uploads for portfolio " + portfolioId);
-		List<S3ObjectSummary> response = storageService.getUploads(null, portfolioId).getObjectSummaries();
-		return ResponseEntity.ok(response);
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		if (portfolioItem == null) throw new PortfolioItemNotFoundException(portfolioId);
+
+		if (canUploadFilesToPortfolio(portfolioItem)) {
+			Page<FileInformationDTO> response = fileInformationService.findAllForPorfolioItem(portfolioId, pageable);
+			return ResponseEntity.ok(response.getContent());
+		} else {
+			log.info("User is not authorized to retrieve uploaded files for portfolio item {}.", portfolioId);
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
-	@GetMapping(path="/{portfolioId}/upload/{keyName:.+}")
+	@GetMapping(path="/{portfolioId}/upload/{id}")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
 	public ResponseEntity getUpload(@PathVariable final Long portfolioId,
-									@PathVariable final String keyName) {
-		log.debug("GET request to get course upload : {}", keyName);
-		InputStream response = storageService.getUpload(null, portfolioId, keyName);
-		if (response != null) {
+									@PathVariable final Long id) {
+		log.debug("GET request to get portfolio upload : {}", id);
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		FileInformationDTO fileInformationDTO = fileInformationService.findOne(id);
+
+		if (portfolioItem == null) throw new PortfolioItemNotFoundException(portfolioId);
+		if (fileInformationDTO == null) throw new FileInformationNotFoundException(id);
+
+		if (canUploadFilesToPortfolio(portfolioItem) && fileInformationDTO.getPortfolioItemId().equals(portfolioId)) {
+			String fileName = fileInformationService.getFileNameFor(id);
+			InputStream response = storageService.getUpload(id);
 			try {
 				byte[] out = org.apache.commons.io.IOUtils.toByteArray(response);
 
 				HttpHeaders responseHeaders = new HttpHeaders();
-				responseHeaders.add("content-disposition", "attachment; filename=" + keyName);
-				// responseHeaders.add("Content-Type", type);
+				responseHeaders.add("content-disposition", "attachment; filename=" + fileName);
 
 				return new ResponseEntity(out, responseHeaders, HttpStatus.OK);
-			} catch (Exception e) {
-				new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
+			} catch (IOException e) {
+				log.error(e.getMessage());
+				log.error("File entity '{}' could not be converted to a Byte Array.", id);
+				return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
 			}
-			return new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
 		} else {
-			return new ResponseEntity ("File Not Found", HttpStatus.NOT_FOUND);
+			log.info("User is not authorized to retrieve uploaded file {}.", id);
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
 		}
 	}
 
-	@DeleteMapping(path="/{portfolioId}/upload/{keyName:.+}")
+	@DeleteMapping(path="/{portfolioId}/upload/{id}")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
 	public ResponseEntity deleteUpload(@PathVariable final Long portfolioId,
-									   @PathVariable final String keyName) {
-		try {
-			storageService.deleteUpload(null, portfolioId, keyName);
-		} catch (Exception e) {
-			return new ResponseEntity("Error deleting upload", HttpStatus.INTERNAL_SERVER_ERROR);
+									   @PathVariable final Long id) {
+		PortfolioItemDTO portfolioItem = portfolioItemService.findOne(portfolioId);
+		FileInformationDTO fileInformationDTO = fileInformationService.findOne(id);
+
+		if (portfolioItem == null) throw new PortfolioItemNotFoundException(portfolioId);
+		if (fileInformationDTO == null) throw new FileInformationNotFoundException(id);
+
+		if (canUploadFilesToPortfolio(portfolioItem) && fileInformationDTO.getPortfolioItemId().equals(portfolioId)) {
+			try {
+				storageService.deleteUpload(id);
+			} catch (Exception e) {
+				log.error(e.getMessage());
+				return new ResponseEntity("Error deleting upload", HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			return new ResponseEntity(HttpStatus.OK);
+		} else {
+			log.info("User is not authorized to delete uploaded file {}.", id);
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
 		}
-		return new ResponseEntity(HttpStatus.OK);
+	}
+
+	private Boolean canUploadFilesToPortfolio(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isAdmin() ||
+			isOrgAdmin(portfolioItem);
+	}
+
+	private Boolean hasCreateUpdateDeleteAuthority(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isAdmin() ||
+			isOrgAdmin(portfolioItem);
+	}
+
+	private Boolean isOrgAdmin(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ORG_ADMIN) &&
+			portfolioItemService.isOrgAdminOfCurrentUser(portfolioItem);
+
+	}
+
+	private Boolean isPortfolioStudent(PortfolioItemDTO portfolioItem) {
+		return SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.STUDENT) &&
+			portfolioItem.getStudentId().equals(userService.getCurrentUser().getId());
 	}
 }
